+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## Fixed
+- Fixed missing parents in consensus
+- Fixed circe mappings for ADTs
 
 ## [v2.9.0] - 2020-06-11
 ## Changed